# swagger/swagger.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:9090
...
paths:
  /api/v1/certification:
    post:
      operationId: createCertification
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificationCreateDto"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CertificationDto"
          description: Certification created
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Create a new certification
      tags:
      - Certification
    put:
      operationId: updateCertification
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificationUpdateDto"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CertificationDto"
          description: Certification updated
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Update an existing certification
      tags:
      - Certification
  /api/v1/certification/{id}:
    delete:
      operationId: deleteCertification
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Certification deleted
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Delete a certification by ID
      tags:
      - Certification
    get:
      operationId: getCertificationById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CertificationDto"
          description: Certification found
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Get certification by ID
      tags:
      - Certification
  /api/v1/company:
    post:
      operationId: createCompany
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyCreateDto"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyDto"
          description: Company record created
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Create a new company record
      tags:
      - Company
    put:
      operationId: updateCompany
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdateDto"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyDto"
          description: Company record updated
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Update an existing company record
      tags:
      - Company
  /api/v1/company/{id}:
    delete:
      operationId: deleteCompany
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Company record deleted
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Delete a company record by ID
      tags:
      - Company
    get:
      operationId: getCompanyById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyDto"
          description: Company record found
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Get company record by ID
      tags:
      - Company
...
components:
  schemas:
    BadRequestErrorResponse:
      type: object
      description: Bad Request Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "400"
        message:
          type: string
          description: Description of the error
          example: Invalid input provided
    CertificationCreateDto:
      type: object
      description: Certification Create Model
      properties:
        credentialUrl:
          type: string
          description: URL providing details about the credential
          maxLength: 255
          minLength: 0
        expirationDate:
          type: string
          format: date-time
          description: Expiration date of the certification
        issueDate:
          type: string
          format: date-time
          description: Issue date of the certification
        issuingOrganization:
          type: string
          description: Issuer of the certification
          maxLength: 255
          minLength: 0
        languageProficiency:
          type: string
          description: Proficiency level in the associated language
          enum:
          - A1
          - A2
          - B1
          - B2
          - C1
          - C2
        name:
          type: string
          description: Certification name
          maxLength: 255
          minLength: 0
        resumeId:
          type: integer
          format: int64
          description: Resume ID
      required:
      - issuingOrganization
      - name
    CertificationDto:
      type: object
      description: Certification Model
      properties:
        credentialUrl:
          type: string
          description: URL with details about the certification
        expirationDate:
          type: string
          format: date-time
          description: Date when the certification expires
        id:
          type: integer
          format: int64
          description: Unique identifier of the certification
        issueDate:
          type: string
          format: date-time
          description: Date when the certification was issued
        issuingOrganization:
          type: string
          description: Organization that issued the certification
        languageProficiency:
          type: string
          description: Proficiency level of the associated language
          enum:
          - A1
          - A2
          - B1
          - B2
          - C1
          - C2
        name:
          type: string
          description: Name of the certification
    CertificationUpdateDto:
      type: object
      description: Certification Update Model
      properties:
        credentialUrl:
          type: string
          description: URL providing details about the credential
          maxLength: 255
          minLength: 0
        expirationDate:
          type: string
          format: date-time
          description: Expiration date of the certification
        id:
          type: integer
          format: int64
          description: Certification ID
        issueDate:
          type: string
          format: date-time
          description: Issue date of the certification
        issuingOrganization:
          type: string
          description: Issuer of the certification
          maxLength: 255
          minLength: 0
        languageProficiency:
          type: string
          description: Proficiency level in the associated language
          enum:
          - A1
          - A2
          - B1
          - B2
          - C1
          - C2
        name:
          type: string
          description: Certification name
          maxLength: 255
          minLength: 0
        resumeId:
          type: integer
          format: int64
          description: Resume ID
    CompanyCreateDto:
      type: object
      description: Company Create Model
      properties:
        name:
          type: string
          description: Name of the company
          maxLength: 255
          minLength: 0
        url:
          type: string
          description: Company URL
          maxLength: 255
          minLength: 0
      required:
      - name
    CompanyDto:
      type: object
      description: Company Model
      properties:
        id:
          type: integer
          format: int64
          description: ID of the company
        name:
          type: string
          description: Name of the company
        url:
          type: string
          description: Company URL
    CompanyUpdateDto:
      type: object
      description: Company Update Model
      properties:
        id:
          type: integer
          format: int64
          description: ID of the company
        name:
          type: string
          description: Name of the company
          maxLength: 255
          minLength: 0
        url:
          type: string
          description: Company URL
          maxLength: 255
          minLength: 0
    ConflictErrorResponse:
      type: object
      description: Conflict Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "409"
        message:
          type: string
          description: Description of the error
          example: Data Conflict
...
    InternalServerErrorResponse:
      type: object
      description: Internal Server Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "500"
        message:
          type: string
          description: Description of the error
          example: Internal Server Error
...
    NotFoundErrorResponse:
      type: object
      description: Not Found Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "404"
        message:
          type: string
          description: Description of the error
          example: Not Found
...